Module Spawn.Pgid

Process group IDs

type t

Representation of the second parameter to setpgid. If a value of this type is provided to spawn, the child will immediately set its pgid accordingly.

val new_process_group : t

Sets the child's pgid to the same as its process id. Equivalent to calling setpgid(0, 0).

val of_pid : int -> t

Raises Invalid_arg if the value is not strictly positive.